Dynamic validation with validator
The following changes comes with this commit:

New Validation property in the mockdata that mirrors the validation object from a api response.
New handler function in FromAgent that runs all validation rules for a question and returns true or false.
New way of handling the withForm property on a question within FormAgent.

These changes does not include error messages for a validation, it was however taken in consideration for supporting error messages.