fix(case-overview): setting timeout before fetching cases from api
Due to varying creation time for a case in the api new active cases was not allways shown in the
list. The case list had to be refreshed in order to update. To fix this a delay have been added
before starting the fetch request, so that the api have time to process everything it has to do on
case creation.