refactor(case-data-converter): new function for replacing parts in a tag replaceTagPart
A new function have been added to the CaseDataConverter helper replaceTagPart. This function takes a
tag, the name of a part in the tag and a replace value for the part then replaces all instances
where the part is defined in singular. This function should only replace a part if the whole part
can be matched with the passed part name. An example part1:part2:part3 say we want to replace part1,
then we pass part1 as the value of the part to be replaced and the value that should replace it for
example replaced. Then the tag should look like replaced:part2:part3